Signature functions moved to rpc_types_compat
