+++ crates/rpc/rpc-types-compat/Cargo.toml
@@ -14,3 +14,6 @@ Compatibility layer for reth-primitives and ethereum RPC types
 reth-primitives.workspace = true
 reth-rpc-types.workspace = true
 reth-rlp.workspace = true
+
+#misc
+serde = { workspace = true, features = ["derive"] }
