+++ crates/rpc/rpc-types-compat/src/lib.rs
@@ -19,3 +19,4 @@ pub mod block;
 pub use block::*;
 pub mod transaction;
 pub use transaction::*;
+
